Implemented bonus points and deleting hunters!
FINALLY! Been looking into the p5.play library for a long time and tried all sorts of things but I could not get it to work. I've come up with a sort of work around to ditch the p5.play library entirely and created my own collision detection, now I can delete hunters! The main mechanics of the game are now in place in this build. I also figured out how to reset the game too.